Move Counter component away from container
